fix(pr-1014): rebase on main + extract auto-compact to satisfy file-size ceiling
- Rebase onto main to pick up updated .filesize-baseline.json and
  .funcsize-baseline.json (PR #1012 bumped baselines for files that
  grew on main after this branch diverged — the stale baselines caused
  6 false-positive violations in CI).

- Extract `maybeAutoCompact()` from query-turn-driver.ts into
  query-turn-driver.auto-compact.ts (353 → 288 lines) so the file
  stays under the 350-line ceiling. The function was already standalone
  with explicit parameters; only the physical location changed.
